+++ src/ska_low_mccs_daq/pydaq/daq_receiver_interface.py
@@ -52,6 +52,14 @@ class DaqReceiver:
 
         _fields_ = [("x", ctypes.c_int32), ("y", ctypes.c_int32)]
 
+    class Diagnostic(ctypes.Structure):
+        """Diagnostic structure definition"""
+
+        _fields_ = [
+            ("ringbuffer_occupancy", ctypes.c_double),
+            ("lost_pushes", ctypes.c_size_t),
+        ]
+
     class ChannelMetadata(ctypes.Structure):
         _fields_ = [
             ("tile", ctypes.c_int16),
@@ -101,6 +109,9 @@ class DaqReceiver:
         ctypes.c_uint32,
         ctypes.c_uint32,
     )
+
+    DIAGNOSTIC_CALLBACK = ctypes.CFUNCTYPE(None, ctypes.POINTER(ctypes.c_void_p))
+
     DYNAMIC_DATA_CALLBACK = ctypes.CFUNCTYPE(
         None,
         ctypes.POINTER(ctypes.c_void_p),
@@ -210,11 +221,39 @@ class DaqReceiver:
         # Pointer to logging function forwarded by low-level DAQ
         self._daq_logging_function = self.LOGGER_CALLBACK(self._logging_callback)
 
+        # Pointer to diagnostic function forwarded by low-level DAQ
+        self._daq_diagnostic_callback = self.DIAGNOSTIC_CALLBACK(
+            self._diagnostic_callback
+        )
+
         # Keep track of which data consumers are running
         self._running_consumers = {}
 
         # Placeholder for continuous and station data to skip first few buffers
         self._buffer_counter = {}
+
+    # --------------------------------------- DIAGNOSTICS ------------------------------------
+
+    def _diagnostic_callback(self, data: ctypes.POINTER) -> None:
+        """Diagnostic callback
+        :param data: Received data
+        """
+        # Extract diagnostic data
+        diagnostic = ctypes.cast(data, ctypes.POINTER(self.Diagnostic)).contents
+        diagnostics = {
+            "ringbuffer_occupancy": diagnostic.ringbuffer_occupancy,
+            "lost_pushes": diagnostic.lost_pushes,
+        }
+
+        # Call external diagnostic callback if defined
+        if self._external_diagnostic_callback is not None:
+            self._external_diagnostic_callback(**diagnostics)
+        elif self._config["logging"]:
+            logging.debug(
+                "Received diagnostic data - occupancy: {}, lost: {}".format(
+                    diagnostic.ringbuffer_occupancy, diagnostic.lost_pushes
+                )
+            )
 
     # --------------------------------------- CONSUMERS --------------------------------------
 
@@ -396,6 +435,7 @@ class DaqReceiver:
                     append=True,
                     data_ptr=values,
                     timestamp=self._timestamps[DaqModes.INTEGRATED_CHANNEL_DATA],
+                    sampling_time=self._sampling_time[DaqModes.INTEGRATED_CHANNEL_DATA],
                     buffer_timestamp=timestamp,
                     tile_id=tile,
                 )
@@ -1447,10 +1487,13 @@ class DaqReceiver:
         self,
         daq_modes: Union[DaqModes, List[DaqModes]],
         callbacks: Optional[Union[Callable, List[Callable]]] = None,
+        diagnostic_callback: Optional[Callable] = None,
     ) -> None:
         """Start acquiring data for specified modes
         :param daq_modes: List of modes to start, should be from DaqModes
-        :param callbacks: List of callbacks, one per mode in daq_modes"""
+        :param callbacks: List of callbacks, one per mode in daq_modes
+        :param diagnostic_callback: Callback for diagnostics
+        """
 
         if type(daq_modes) != list:
             daq_modes = [daq_modes]
@@ -1465,6 +1508,8 @@ class DaqReceiver:
                 "Number of callback should match number of daq_modes. Ignoring callbacks."
             )
             callbacks = []
+
+        self._external_diagnostic_callback = diagnostic_callback
 
         # Check all modes
         for i, mode in enumerate(daq_modes):
@@ -1655,7 +1700,6 @@ class DaqReceiver:
 
     def get_configuration(self) -> Dict[str, Any]:
         """Return configuration dictionary"""
-        logging.warning(f"the config is {self._config}")
         return self._config
 
     @staticmethod
@@ -1847,13 +1891,18 @@ class DaqReceiver:
         self._daq_library.initialiseConsumer.restype = ctypes.c_int
 
         # Define startConsumer function
-        self._daq_library.startConsumer.argtypes = [ctypes.c_char_p, self.DATA_CALLBACK]
+        self._daq_library.startConsumer.argtypes = [
+            ctypes.c_char_p,
+            self.DATA_CALLBACK,
+            self.DIAGNOSTIC_CALLBACK,
+        ]
         self._daq_library.startConsumer.restype = ctypes.c_int
 
         # Define startConsumer function
         self._daq_library.startConsumerDynamic.argtypes = [
             ctypes.c_char_p,
             self.DYNAMIC_DATA_CALLBACK,
+            self.DIAGNOSTIC_CALLBACK,
         ]
         self._daq_library.startConsumerDynamic.restype = ctypes.c_int
 
@@ -1949,9 +1998,13 @@ class DaqReceiver:
 
         # Start consumer
         if dynamic_callback:
-            res = self._daq_library.startConsumerDynamic(consumer, callback)
+            res = self._daq_library.startConsumerDynamic(
+                consumer, callback, self._daq_diagnostic_callback
+            )
         else:
-            res = self._daq_library.startConsumer(consumer, callback)
+            res = self._daq_library.startConsumer(
+                consumer, callback, self._daq_diagnostic_callback
+            )
         if res != self.Result.Success.value:
             return self.Result.Failure